backend/requirements.txt
certifi
python-dotenv
google-generativeai
flask
flask-cors
pytest
requests
